rename playlist to player.playqueue
