+++ app/src/main/java/org/schabi/newpipe/database/stream/model/StreamEntity.java
@@ -9,7 +9,7 @@ import android.arch.persistence.room.PrimaryKey;
 import org.schabi.newpipe.extractor.stream.StreamInfo;
 import org.schabi.newpipe.extractor.stream.StreamInfoItem;
 import org.schabi.newpipe.extractor.stream.StreamType;
-import org.schabi.newpipe.playlist.PlayQueueItem;
+import org.schabi.newpipe.playqueue.PlayQueueItem;
 import org.schabi.newpipe.util.Constants;
 
 import java.io.Serializable;
