+++ app/src/main/java/org/schabi/newpipe/fragments/detail/VideoDetailFragment.java
@@ -71,8 +71,8 @@ import org.schabi.newpipe.player.MainVideoPlayer;
 import org.schabi.newpipe.player.PopupVideoPlayer;
 import org.schabi.newpipe.player.helper.PlayerHelper;
 import org.schabi.newpipe.player.old.PlayVideoActivity;
-import org.schabi.newpipe.playlist.PlayQueue;
-import org.schabi.newpipe.playlist.SinglePlayQueue;
+import org.schabi.newpipe.playqueue.PlayQueue;
+import org.schabi.newpipe.playqueue.SinglePlayQueue;
 import org.schabi.newpipe.report.UserAction;
 import org.schabi.newpipe.util.Constants;
 import org.schabi.newpipe.util.ExtractorHelper;
