+++ app/src/main/java/org/schabi/newpipe/fragments/list/BaseListFragment.java
@@ -24,7 +24,7 @@ import org.schabi.newpipe.fragments.local.dialog.PlaylistAppendDialog;
 import org.schabi.newpipe.local.history.HistoryInfoItem;
 import org.schabi.newpipe.info_list.InfoItemDialog;
 import org.schabi.newpipe.info_list.InfoListAdapter;
-import org.schabi.newpipe.playlist.SinglePlayQueue;
+import org.schabi.newpipe.playqueue.SinglePlayQueue;
 import org.schabi.newpipe.util.NavigationHelper;
 import org.schabi.newpipe.util.OnClickGesture;
 import org.schabi.newpipe.util.StateSaver;
