+++ app/src/main/java/org/schabi/newpipe/fragments/list/channel/ChannelFragment.java
@@ -36,9 +36,9 @@ import org.schabi.newpipe.extractor.stream.StreamInfoItem;
 import org.schabi.newpipe.fragments.list.BaseListInfoFragment;
 import org.schabi.newpipe.fragments.local.dialog.PlaylistAppendDialog;
 import org.schabi.newpipe.info_list.InfoItemDialog;
-import org.schabi.newpipe.playlist.ChannelPlayQueue;
-import org.schabi.newpipe.playlist.PlayQueue;
-import org.schabi.newpipe.playlist.SinglePlayQueue;
+import org.schabi.newpipe.playqueue.ChannelPlayQueue;
+import org.schabi.newpipe.playqueue.PlayQueue;
+import org.schabi.newpipe.playqueue.SinglePlayQueue;
 import org.schabi.newpipe.report.UserAction;
 import org.schabi.newpipe.local.subscription.SubscriptionService;
 import org.schabi.newpipe.util.AnimationUtils;
