+++ app/src/main/java/org/schabi/newpipe/player/MainVideoPlayer.java
@@ -62,10 +62,10 @@ import org.schabi.newpipe.extractor.stream.VideoStream;
 import org.schabi.newpipe.fragments.OnScrollBelowItemsListener;
 import org.schabi.newpipe.player.helper.PlaybackParameterDialog;
 import org.schabi.newpipe.player.helper.PlayerHelper;
-import org.schabi.newpipe.playlist.PlayQueueItem;
-import org.schabi.newpipe.playlist.PlayQueueItemBuilder;
-import org.schabi.newpipe.playlist.PlayQueueItemHolder;
-import org.schabi.newpipe.playlist.PlayQueueItemTouchCallback;
+import org.schabi.newpipe.playqueue.PlayQueueItem;
+import org.schabi.newpipe.playqueue.PlayQueueItemBuilder;
+import org.schabi.newpipe.playqueue.PlayQueueItemHolder;
+import org.schabi.newpipe.playqueue.PlayQueueItemTouchCallback;
 import org.schabi.newpipe.util.AnimationUtils;
 import org.schabi.newpipe.util.ListHelper;
 import org.schabi.newpipe.util.NavigationHelper;
