+++ app/src/main/java/org/schabi/newpipe/player/ServicePlayerActivity.java
@@ -32,11 +32,11 @@ import org.schabi.newpipe.fragments.OnScrollBelowItemsListener;
 import org.schabi.newpipe.fragments.local.dialog.PlaylistAppendDialog;
 import org.schabi.newpipe.player.event.PlayerEventListener;
 import org.schabi.newpipe.player.helper.PlaybackParameterDialog;
-import org.schabi.newpipe.playlist.PlayQueueAdapter;
-import org.schabi.newpipe.playlist.PlayQueueItem;
-import org.schabi.newpipe.playlist.PlayQueueItemBuilder;
-import org.schabi.newpipe.playlist.PlayQueueItemHolder;
-import org.schabi.newpipe.playlist.PlayQueueItemTouchCallback;
+import org.schabi.newpipe.playqueue.PlayQueueAdapter;
+import org.schabi.newpipe.playqueue.PlayQueueItem;
+import org.schabi.newpipe.playqueue.PlayQueueItemBuilder;
+import org.schabi.newpipe.playqueue.PlayQueueItemHolder;
+import org.schabi.newpipe.playqueue.PlayQueueItemTouchCallback;
 import org.schabi.newpipe.util.Localization;
 import org.schabi.newpipe.util.NavigationHelper;
 import org.schabi.newpipe.util.ThemeHelper;
