+++ app/src/main/java/org/schabi/newpipe/player/mediasource/LoadedMediaSource.java
@@ -7,7 +7,7 @@ import com.google.android.exoplayer2.source.MediaPeriod;
 import com.google.android.exoplayer2.source.MediaSource;
 import com.google.android.exoplayer2.upstream.Allocator;
 
-import org.schabi.newpipe.playlist.PlayQueueItem;
+import org.schabi.newpipe.playqueue.PlayQueueItem;
 
 import java.io.IOException;
