+++ app/src/main/java/org/schabi/newpipe/player/playback/BasePlayerMediaSession.java
@@ -5,7 +5,7 @@ import android.support.v4.media.MediaDescriptionCompat;
 
 import org.schabi.newpipe.player.BasePlayer;
 import org.schabi.newpipe.player.mediasession.MediaSessionCallback;
-import org.schabi.newpipe.playlist.PlayQueueItem;
+import org.schabi.newpipe.playqueue.PlayQueueItem;
 
 public class BasePlayerMediaSession implements MediaSessionCallback {
     private BasePlayer player;
