+++ app/src/main/java/org/schabi/newpipe/player/playback/PlaybackListener.java
@@ -6,7 +6,7 @@ import android.support.annotation.Nullable;
 import com.google.android.exoplayer2.source.MediaSource;
 
 import org.schabi.newpipe.extractor.stream.StreamInfo;
-import org.schabi.newpipe.playlist.PlayQueueItem;
+import org.schabi.newpipe.playqueue.PlayQueueItem;
 
 import java.util.List;
