+++ app/src/main/java/org/schabi/newpipe/player/playqueue/PlayQueueItemHolder.java
@@ -1,4 +1,4 @@
-package org.schabi.newpipe.playlist;
+package org.schabi.newpipe.playqueue;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
