+++ app/src/main/java/org/schabi/newpipe/player/playqueue/StatisticsPlaylistFragment.java
@@ -1,4 +1,4 @@
-package org.schabi.newpipe.playlist;
+package org.schabi.newpipe.playqueue;
 
 import android.app.Activity;
 import android.content.Context;
