+++ app/src/main/java/org/schabi/newpipe/player/playqueue/events/RecoveryEvent.java
@@ -1,4 +1,4 @@
-package org.schabi.newpipe.playlist.events;
+package org.schabi.newpipe.playqueue.events;
 
 
 public class RecoveryEvent implements PlayQueueEvent {
